% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lavaan_estimates.R
\name{get_sigma_e}
\alias{get_sigma_e}
\title{Title}
\usage{
get_sigma_e(lavaan_fit)
}
\arguments{
\item{lavaan_fit}{a fitted lavaan model}
}
\value{
correlation matrix for error from fitted model
}
\description{
Title
}
