% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_estimates.R
\name{import_estimates_stata}
\alias{import_estimates_stata}
\title{Import esitmates from stata}
\usage{
import_estimates_stata(e_b, e_b_rownames, e_V)
}
\arguments{
\item{e_b}{A vector of parameter estimates from a sem model in stata}

\item{e_b_rownames}{The (long) Stata names of the paramter estimates. Will have format "outcome:parameter".}

\item{e_V}{The variance/covariance matrix for the parameters}
}
\value{
A list containing the parameter estimates,
their covariance matrix, the covariance matrix for the errors,
and counts for the number of paramaters and covariates in the model
}
\description{
Import esitmates from stata
}
\examples{
load("data/e_V.rda")
load("data/e_b.rda")
import_estimates_stata(e_b = e_b[[1]], e_b_rownames = e_b[[2]], e_V = e_V)
}
