% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_predictions}
\alias{plot_predictions}
\title{plot_predictions}
\usage{
plot_predictions(predictions)
}
\arguments{
\item{predictions}{a tibble of predictions created by mopi::predict_weighted_lavaan}
}
\description{
plot_predictions
}
