% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_weighted_lavaan.R
\name{predict_weighted_lavaan}
\alias{predict_weighted_lavaan}
\title{predict_weighted_lavaan}
\usage{
predict_weighted_lavaan(new_x, weights, lavaan_model, predict_outcomes = TRUE)
}
\arguments{
\item{new_x}{vector of covariates to predict}

\item{weights}{weights to apply to each outcome}

\item{lavaan_model}{fitted lavaan mdoel}

\item{predict_outcomes}{TRUE/FALSE: Indicates whether predictions of individual outcomes are included or only weighted prediction}
}
\value{

}
\description{
predict_weighted_lavaan
}
